+++ packages/ckeditor5-clipboard/src/dragdropblocktoolbar.ts
@@ -1,0 +1,145 @@
+/**
+ * @license Copyright (c) 2003-2023, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
+ */
+
+/**
+ * @module clipboard/dragdropblocktoolbar
+ */
+
+import { Plugin } from '@ckeditor/ckeditor5-core';
+
+import {
+	env,
+	global,
+	DomEmitterMixin,
+	type ObservableChangeEvent,
+	type DomEmitter
+} from '@ckeditor/ckeditor5-utils';
+
+import type { BlockToolbar } from '@ckeditor/ckeditor5-ui';
+
+import ClipboardObserver from './clipboardobserver';
+
+/**
+ * TODO
+ */
+export default class DragDropBlockToolbar extends Plugin {
+	/**
+	 * TODO
+	 */
+	private _isBlockDragging = false;
+
+	/**
+	 * TODO
+	 */
+	private _domEmitter: DomEmitter = new ( DomEmitterMixin() )();
+
+	/**
+	 * @inheritDoc
+	 */
+	public static get pluginName(): 'DragDropBlockToolbar' {
+		return 'DragDropBlockToolbar';
+	}
+
+	/**
+	 * @inheritDoc
+	 */
+	public init(): void {
+		const editor = this.editor;
+
+		this.listenTo<ObservableChangeEvent<boolean>>( editor, 'change:isReadOnly', ( evt, name, isReadOnly ) => {
+			if ( isReadOnly ) {
+				this.forceDisabled( 'readOnlyMode' );
+				this._isBlockDragging = false;
+			} else {
+				this.clearForceDisabled( 'readOnlyMode' );
+			}
+		} );
+
+		if ( env.isAndroid ) {
+			this.forceDisabled( 'noAndroidSupport' );
+		}
+
+		if ( editor.plugins.has( 'BlockToolbar' ) ) {
+			const blockToolbar: BlockToolbar = editor.plugins.get( 'BlockToolbar' );
+			const element = blockToolbar.buttonView.element!;
+
+			element.setAttribute( 'draggable', 'true' );
+
+			this._domEmitter.listenTo( element, 'dragstart', ( evt, data ) => this._handleBlockDragStart( data ) );
+			this._domEmitter.listenTo( global.document, 'dragover', ( evt, data ) => this._handleBlockDragging( data ) );
+			this._domEmitter.listenTo( global.document, 'drop', ( evt, data ) => this._handleBlockDragging( data ) );
+			this._domEmitter.listenTo( global.document, 'dragend', () => this._handleBlockDragEnd(), { useCapture: true } );
+		}
+	}
+
+	/**
+	 * @inheritDoc
+	 */
+	public override destroy(): void {
+		this._domEmitter.stopListening();
+
+		return super.destroy();
+	}
+
+	/**
+	 * TODO
+	 */
+	private _handleBlockDragStart( domEvent: DragEvent ): void {
+		if ( !this.isEnabled ) {
+			return;
+		}
+
+		const model = this.editor.model;
+		const selection = model.document.selection;
+
+		const blocks = Array.from( selection.getSelectedBlocks() );
+		const draggedRange = model.createRange(
+			model.createPositionBefore( blocks[ 0 ] ),
+			model.createPositionAfter( blocks[ blocks.length - 1 ] )
+		);
+
+		model.change( writer => writer.setSelection( draggedRange ) );
+
+		this._isBlockDragging = true;
+
+		this.editor.editing.view.getObserver( ClipboardObserver )!.onDomEvent( domEvent );
+	}
+
+	/**
+	 * TODO
+	 */
+	private _handleBlockDragging( domEvent: DragEvent ): void {
+		if ( !this.isEnabled ) {
+			return;
+		}
+
+		const clientX = domEvent.clientX + 100;
+		const clientY = domEvent.clientY;
+		const target = document.elementFromPoint( clientX, clientY );
+
+		if ( !target || !target.closest( '.ck-editor__editable' ) || !this._isBlockDragging ) {
+			return;
+		}
+
+		this.editor.editing.view.getObserver( ClipboardObserver )!.onDomEvent( {
+			...domEvent,
+			type: domEvent.type,
+			dataTransfer: domEvent.dataTransfer,
+			target,
+			clientX,
+			clientY,
+			preventDefault: () => domEvent.preventDefault(),
+			stopPropagation: () => domEvent.stopPropagation()
+		} );
+	}
+
+	/**
+	 * TODO
+	 */
+	private _handleBlockDragEnd(): void {
+		this._isBlockDragging = false;
+	}
+}
+
