+++ packages/ckeditor5-theme-lark/theme/ckeditor5-ui/components/toolbar/toolbar.css
@@ -106,6 +106,10 @@
 		}
 	}
 
+	&:focus {
+		outline: none;
+	}
+
 	@nest .ck-toolbar-container & {
 		border: 0;
 	}
